+++ database/migrations/2019_03_30_104225_create_table_buys.php
@@ -15,9 +15,9 @@ class CreateTableBuys extends Migration
     {
         Schema::create('Buys', function (Blueprint $table) {
             $table->increments('id');
-            $table->mediumInteger('buyer_id');
-            $table->mediumInteger('pro_id');
-            $table->smallInteger('shop_id');
+            $table->integer('buyer_id');
+            $table->integer('pro_id');
+            $table->integer('shop_id');
             $table->tinyInteger('num_buy')->comment('تعداد خرید');
             $table->mediumInteger('price_pro')->comment('قیمت واحد کالا در زمان درج ستون ، ممکن است بعد از خرید کالا قیمت محصول تغییر کند ، نباید مقدار این ستون تغییر کند');
             $table->integer('date_up');
